refactor: synchronise structure of both resource pages
