+++ frontend/src/features/resources/EditableResourceListItem.tsx
@@ -1,0 +1,50 @@
+import type React from "react";
+import ResourceNameField from "./components/ResourceNameField";
+import { isTemp } from "./util";
+import { XIcon } from "lucide-react";
+import { useResources } from "./useResources";
+import type { UseMutationResult } from "@tanstack/react-query";
+
+interface ResourceListItemProps {
+  resource: {
+    _id: string;
+    name: string;
+  };
+  setSelectedResourceId: React.Dispatch<React.SetStateAction<string | null>>;
+}
+
+function EditableResourceListItem({
+  resource,
+  setSelectedResourceId,
+}: ResourceListItemProps) {
+  const { renameResourceMutation, deleteResourceMutation } = useResources();
+
+  return (
+    <ResourceNameField
+      resource={resource}
+      disabled={isTemp(resource) as boolean}
+      onSelect={() => setSelectedResourceId(resource._id)}
+      onRename={(name: string) =>
+        (renameResourceMutation as UseMutationResult).mutate({
+          resourceId: resource._id,
+          name,
+        })
+      }
+      actions={
+        <button
+          className="btn btn-phantom btn-xs px-0"
+          onClick={(e) => {
+            e.stopPropagation();
+            (deleteResourceMutation as UseMutationResult).mutate(resource._id);
+          }}
+          title="Delete file"
+          disabled={isTemp(resource) as boolean}
+        >
+          <XIcon size={16} />
+        </button>
+      }
+    />
+  );
+}
+
+export default EditableResourceListItem;
